SimPe 0.51e (QA) ---------------------------------------------------------     + Added the Folder for the CHristmas Objects to the default FileTable (Thanks to Carrigon)     + Fixed the Problem with the Finder Pinhead did report     + Added the "Chees" Aspiration     + Sim Listing on the "Sim Relationship" Tab was extended. NOw it sorts like this: 1. All sims with a Relationship, 2. All Usermade Sims, 3. All NPCs, 4. All Townies     + Fixed some more Memory Leaks     + Fixed the spelling Error in the TXTM Plugin reported by SnowStorm     + Fixed Visibility of the "EP-Ready 2 Version" Checkboy and the "assign hash" Button (as purposed by Numenor)

+++ trunk/fullsimpe/SimPe Helper/Registry.cs
@@ -883,7 +883,7 @@ namespace SimPe
 			{
 				XmlRegistryKey  rkf = xrk.CreateSubKey("Settings");
 				object o = rkf.GetValue("BigPackageResourceCount", 1000);
-				return Convert.ToInt16(o);
+				return Convert.ToInt32(o);
 			}
 			set
 			{
